# frontend/src/api/generated/openapi.json
{
  "components": {
    "schemas": {
...
      "DeleteNoteResponse": {
        "description": "删除成功只回执被删的笔记标识。",
        "properties": {
          "deleted": {
            "title": "Deleted",
            "type": "string"
          }
        },
        "required": [
          "deleted"
        ],
        "title": "DeleteNoteResponse",
        "type": "object"
      },
      "FailureDetail": {
        "description": "业务失败的 detail：各路由统一用 `{\"message\": ...}` 包装。",
        "properties": {
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "message"
        ],
        "title": "FailureDetail",
        "type": "object"
      },
      "FailurePayload": {
        "description": "业务失败响应：哪个状态码适用仍由各端点自己声明（`responses`）。",
        "properties": {
          "detail": {
            "$ref": "#/components/schemas/FailureDetail"
          }
        },
        "required": [
          "detail"
        ],
        "title": "FailurePayload",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "NotePayload": {
        "description": "个股笔记：只含正文与最近保存时间，不暴露修订历史。",
        "properties": {
          "body": {
            "title": "Body",
            "type": "string"
          },
          "noteId": {
            "title": "Noteid",
            "type": "string"
          },
          "securityId": {
            "title": "Securityid",
            "type": "string"
          },
          "updatedAt": {
            "title": "Updatedat",
            "type": "string"
          }
        },
        "required": [
          "noteId",
          "securityId",
          "body",
          "updatedAt"
        ],
        "title": "NotePayload",
        "type": "object"
      },
      "NoteStreamPayload": {
        "description": "某股票（以证券为身份）跨导入日期与入口共用的完整笔记流。",
        "properties": {
          "notes": {
            "items": {
              "$ref": "#/components/schemas/NotePayload"
            },
            "title": "Notes",
            "type": "array"
          },
          "securityId": {
            "title": "Securityid",
            "type": "string"
          }
        },
        "required": [
          "securityId",
          "notes"
        ],
        "title": "NoteStreamPayload",
        "type": "object"
      },
      "NoteWrite": {
        "description": "写笔记的请求体：只认识正文，多余字段忽略。\n\n正文刻意不声明成 `str`：迁移前是 `str(payload.get(\"body\") or \"\")`，缺失、null、\n空串与非字符串都先落到字符串再交给领域校验，所以畸形输入的错误是 400／404 而不是\n结构校验 422。若将来要收紧契约，应作为一次单列的行为变化，而不是 DTO 迁移的副作用。",
        "properties": {
          "body": {
            "title": "Body"
          }
        },
        "title": "NoteWrite",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "DailyScreen Lite",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/notes/securities/{security_id}": {
      "get": {
        "description": "该股票跨导入日期与入口共用的完整笔记流。",
        "operationId": "stream_for_security_api_notes_securities__security_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "security_id",
            "required": true,
            "schema": {
              "title": "Security Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NoteStreamPayload"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FailurePayload"
                }
              }
            },
            "description": "股票或笔记不存在"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Stream For Security",
        "tags": [
          "notes"
        ]
      },
      "post": {
        "operationId": "create_for_security_api_notes_securities__security_id__post",
        "parameters": [
          {
            "in": "path",
            "name": "security_id",
            "required": true,
            "schema": {
              "title": "Security Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NoteWrite"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotePayload"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FailurePayload"
                }
              }
            },
            "description": "笔记内容不合规"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FailurePayload"
                }
              }
            },
            "description": "股票或笔记不存在"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create For Security",
        "tags": [
          "notes"
        ]
      }
    },
    "/api/notes/{note_id}": {
      "delete": {
        "operationId": "delete_api_notes__note_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "note_id",
            "required": true,
            "schema": {
              "title": "Note Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteNoteResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FailurePayload"
                }
              }
            },
            "description": "股票或笔记不存在"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete",
        "tags": [
          "notes"
        ]
      },
      "patch": {
        "description": "覆盖正文与保存时间，不保留旧版本。",
        "operationId": "update_api_notes__note_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "note_id",
            "required": true,
            "schema": {
              "title": "Note Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NoteWrite"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotePayload"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FailurePayload"
                }
              }
            },
            "description": "笔记内容不合规"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FailurePayload"
                }
              }
            },
            "description": "股票或笔记不存在"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update",
        "tags": [
          "notes"
        ]
      }
    },
...
  }
}